Refactor: 반복되는 스타일을 apply 설정 재사용성 향상을 위한 apply 설정 Resolves:#6
