+++ components/Buttons/BackButton.tsx
@@ -1,13 +1,14 @@
 "use client";
-import { useRouter } from "next/navigation";
 import React from "react";
+import { useRouter } from "next/navigation";
+import styles from "./Buttons.module.css";
 
 const BackButton = () => {
   const router = useRouter();
   return (
     <button
       type="button"
-      className="w-16 h-16 pt-4 bg-custom-blue rounded-r-lg text-white font-bold bg-arrowTop bg-no-repeat bg-[center_top]"
+      className={`rounded-r-lg ${styles.moveBtn} bg-arrowLeft`}
       onClick={() => {
         router.back();
       }}
